pyproject.toml
[tool.poetry]
name = "dio-bank"
version = "0.1.0"
description = ""
authors = ["Ender Guaina <ender.guaina@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
flask = "^3.1.3"
sqlalchemy = "^2.0.49"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.1.0"
flask-jwt-extended = "^4.7.1"
pytest-mock = "^3.15.1"
psycopg2-binary = "^2.9.12"
gunicorn = "^23.0.0"
flask-bcrypt = "^1.0.1"
flask-marshmallow = "^1.5.0"
marshmallow-sqlalchemy = "^1.5.0"
apispec-webframeworks = "^1.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.3"

[tool.pytest.ini_options]
addopts = "-ra -vvv"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"